Action Buttons Save, Save As and more
App shows meme from list of 10 downloaded memes from r/memes/hot. User could switch between downloaded memes, save the choosen one. App's GUI show name of subreddit, title of the post, number of the post in the list of downloaded memes and shows URL instead of image if image is not included in the post.